Move schema tools to space plugin
